chore(firstMile): move Finish step into shared steps directory
